+++ .github/workflows/demo.yml
@@ -27,6 +27,7 @@ jobs:
           artifact_path: 'debian:9'
           output_sbom_path: 'sbom.json'
           output_inspector_scan_path: 'inspector_scan.json'
+          output_inspector_scan_path_csv: 'inspector_scan.csv'
 
       - name: Demonstrate SBOM Output
         run: python3 scripts/decode_action_output.py ${{ steps.inspector.outputs.artifact_sbom }}
@@ -42,4 +43,5 @@ jobs:
           path: |
             sbom.json
             inspector_scan.json
+            inspector_scan.csv
