+++ action.yml
@@ -18,9 +18,14 @@ inputs:
     default: '/tmp/sbom.json'
 
   output_inspector_scan_path:
-    description: "The destination file path for Inspector's vulnerability scan."
+    description: "The destination file path for Inspector's vulnerability scan (JSON format)."
     required: False
     default: '/tmp/inspector_scan.json'
+
+  output_inspector_scan_path_csv:
+    description: "The destination file path for Inspector's vulnerability scan (CSV format)."
+    required: False
+    default: '/tmp/inspector_scan.csv'
 
 
   sbomgen_version:
@@ -57,10 +62,13 @@ inputs:
 
 outputs:
   artifact_sbom:
-    description: 'TBD'
+    description: "The filepath to the artifact SBOM."
 
   inspector_scan_results:
-    description: 'TBD'
+    description: "The filepath to the Inspector vulnerability scan in JSON format."
+
+  inspector_scan_results_csv:
+    description: "The filepath to the Inspector vulnerability scan in CSV format."
 
 runs:
   using: 'docker'
@@ -70,6 +78,7 @@ runs:
     - --artifact-path=${{ inputs.artifact_path }}
     - --out-sbom=${{ inputs.output_sbom_path}}
     - --out-scan=${{ inputs.output_inspector_scan_path }}
+    - --out-scan-csv=${{ inputs.output_inspector_scan_path_csv }}
     - --sbomgen-version=${{ inputs.sbomgen_version }}
     - --thresholds
     - --critical=${{ inputs.critical_threshold }}
