+++ entrypoint/entrypoint/orchestrator.py
@@ -1,3 +1,4 @@
+from entrypoint import converter
 from entrypoint import installer
 from entrypoint import executor
 
@@ -319,6 +320,23 @@ def execute(args) -> int:
                                        lows,
                                        others)
     print(findings)
+
+    # create CSV output
+    csv_output = {}
+    with open(args.out_scan, "r") as f:
+        inspector_scan = json.load(f)
+        vulns = converter.vulns_to_obj(inspector_scan)
+        if vulns:
+            csv_output = converter.to_csv(vulns, artifact_name=args.artifact_path,
+                                          artifact_type=args.artifact_type,
+                                          criticals=criticals,
+                                          highs=highs,
+                                          mediums=mediums,
+                                          lows=lows,
+                                          others=others)
+
+            with open(args.out_scan_csv, "w") as f:
+                f.write(csv_output)
 
     is_exceeded = exceeds_threshold(criticals, args.critical,
                                     highs, args.high,
