+++ entrypoint/tests/test_csv.py
@@ -2,7 +2,7 @@ import json
 import os
 import unittest
 
-from entrypoint import inspector_scan_converter
+from entrypoint import converter
 
 
 class TestCSV(unittest.TestCase):
@@ -17,32 +17,27 @@ class TestCSV(unittest.TestCase):
             with open(path, 'r') as f:
                 inspector_scan_json = json.load(f)
 
-            vulns = inspector_scan_converter.vulns_to_obj(inspector_scan_json)
-            if vulns == None:
+            vulns = converter.vulns_to_obj(inspector_scan_json)
+            if not vulns:
                 continue
 
             for v in vulns:
-                self.assertTrue(["Vulnerability ID"] != "")
-                self.assertTrue(v["Severity"] != "")
-                self.assertTrue(v["Published"] != "")
-                self.assertTrue(v["Modified"] != "")
-                self.assertTrue(v["Description"] != "")
-                self.assertTrue(v["Package Installed Version"] != "")
-                self.assertTrue(v["Package Fixed Version"] != "")
-                self.assertTrue(v["Package Path"] != "")
-                self.assertTrue(v["EPSS Score"] != "")
-                self.assertTrue(v["Exploit Available"] != "")
-                self.assertTrue(v["Exploit Last Seen"] != "")
-                self.assertTrue(v["CWEs"] != "")
+                self.assertTrue(v.vuln_id != "")
+                self.assertTrue(v.severity != "")
+                self.assertTrue(v.published != "")
+                self.assertTrue(v.modified != "")
+                self.assertTrue(v.description != "")
+                self.assertTrue(v.installed_ver != "")
+                self.assertTrue(v.fixed_ver != "")
+                self.assertTrue(v.pkg_path != "")
+                self.assertTrue(v.epss_score != "")
+                self.assertTrue(v.exploit_available != "")
+                self.assertTrue(v.exploit_last_seen != "")
+                self.assertTrue(v.cwes != "")
 
-            as_csv = inspector_scan_converter.to_csv(vulns)
+            as_csv = converter.to_csv(vulns)
             self.assertIsNotNone(as_csv)
             self.assertTrue(as_csv != "")
-
-            f = open("/tmp/csv", "w")
-            f.write(as_csv)
-            f.close()
-            print("")
 
 
 if __name__ == "__main__":
